requirements.txt
requests==2.21.0
urllib3==1.24.3
PyYAML==6.0
